+++ rosen/test/render_service/render_service_base/fuzztest/platform/ohos/rsrenderserviceclient_fuzzer/rsrenderserviceclient_fuzzer.cpp
@@ -334,8 +334,8 @@ bool DoTakeSurfaceCapture(const uint8_t* data, size_t size)
     uint8_t type = GetData<uint8_t>();
     captureConfig.captureType = (SurfaceCaptureType)type;
     captureConfig.isSync = GetData<bool>();
-    uint16_t listSize = GetData<uint16_t>();
-    for (uint16_t i = 0; i < listSize; ++i) {
+    uint8_t listSize = GetData<uint8_t>();
+    for (uint8_t i = 0; i < listSize; ++i) {
         uint64_t nodeid = GetData<uint64_t>();
         captureConfig.blackList.push_back(nodeid);
     }
